+++ src/main/java/es/jose/biblioteca/repositories/ColeccionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
  * 
  * @author joseb85
  */
-public interface ColeccionRepository extends PagingAndSortingRepository<Coleccion, Integer>{
+public interface ColeccionRepository extends PagingAndSortingRepository<Coleccion, Long>{
 
 }
